Refactoring prototype + LICENSE
