+++ README.md
@@ -1,0 +1,10 @@
+# PAGE XML Draw
+
+A CLI application that enables quick visualization of [PAGE-XML](https://github.com/PRImA-Research-Lab/PAGE-XML) files using python.
+
+* adicionar imagem exemplo
+
+## Usage
+
+* guia de instalação
+* mostrar opções de parâmetro
